GH-33 Refactor files names
