+++ src/components/ApplicationStatus/ApplicationStatus.view.tsx
@@ -1,5 +1,5 @@
 import { ApplicationStatus as ApplicationStatusEnum } from '~/@types/common';
-import { styles } from './styles.css';
+import { styles } from './ApplicationStatus.styles.css';
 import theme from '~/theme';
 import useI18n from '~/hooks/useI18n';
