+++ src/components/ApplicationTile/ApplicationTile.styles.css.ts
