+++ src/components/ApplicationTile/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ApplicationTile.view';
