+++ src/components/EmptyState/index.ts
@@ -1,1 +1,1 @@
-export { default } from './EmptyState';
+export { default } from './EmptyState.view';
