+++ src/components/Link/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Link.view';
