+++ src/components/Loader/Loader.view.tsx
