+++ src/components/Loader/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Loader';
+export { default } from './Loader.view';
