+++ src/components/Text/Text.view.tsx
@@ -1,4 +1,4 @@
-import { styles } from './styles.css';
+import { styles } from './Text.styles.css';
 import classNames from 'classnames';
 import { ReactNode } from 'react';
