+++ src/hooks/useGenerator/useGenerator.utils.ts
