+++ src/pages/Generator/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Generator';
+export { default } from './Generator.view';
